requirements.txt
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
ecdsa==0.18.0
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
numpy==1.24.3
pandas==2.0.1
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
